requirements.txt
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
better-profanity==0.7.0
bleach==6.0.0
blis==0.7.11
cached-property==1.5.2
catalogue==1.0.2
certifi==2017.7.27.1
chardet==3.0.4
charset-normalizer==2.0.7
cloudinary==1.28.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.5.0
distlib==0.3.7
Django==4.2.7
django-admin-sortable2==2.1.9
django-allauth==0.33.0
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.0
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-summernote==0.8.20.0
en-core-web-sm==2.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
fonttools==4.42.1
googletrans==3.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
joblib==1.3.2
JsonForm==0.0.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
JsonSir==0.0.2
jupyter-client==8.3.1
jupyter-core==5.3.1
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
more-itertools==8.14.0
murmurhash==1.0.10
nest-asyncio==1.5.8
numpy==1.26.0
oauthlib==2.0.2
openpyxl==3.1.2
ordered-set==3.1.1
ordered-set-stubs==0.1.3
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
plac==1.1.3
platformdirs==3.10.0
playwright==1.38.0
poetry-version==0.1.5
preshed==3.0.9
profanity-check==1.0.3
profanity-filter==1.3.3
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pydantic==1.10.13
pyee==9.0.4
pygame==2.5.2
pygame-ce==2.3.2
pygame-gui==0.6.9
Pygments==2.16.1
pyparsing==3.1.1
pypdf2==3.0.1
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
python-i18n==0.3.9
python3-openid==3.1.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
redis==3.5.3
referencing==0.30.2
requests==2.18.4
requests-oauthlib==0.8.0
Resource==0.2.1
rfc3986==1.5.0
rpds-py==0.10.6
ruamel.yaml==0.15.100
scikit-learn==1.3.2
scipy==1.11.2
seaborn==0.13.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.5
spacy==2.3.9
sqlparse==0.4.2
srsly==1.0.7
stack-data==0.6.2
static3==0.7.0
thinc==7.4.6
threadpoolctl==3.2.0
tomlkit==0.5.11
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.22
virtualenv==20.24.5
wasabi==0.10.1
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.17.0
